+++ frontend/src/webapi.ts
@@ -55,6 +55,7 @@ interface Run {
         name: string;
         pdb_count: number;
         trajectory_count?: number;
+        target_count?: number;
         primary_score_stats?: PrimaryScoreStats;
         results_file: string;
         merged_count?: number;